Store Promotions: Move list state up a level
This moves the list page states up a level to make room for edit states.
