+++ client/extensions/woocommerce/state/selectors/test/promotions.js
@@ -21,9 +21,11 @@ describe( 'promotions', () => {
 			woocommerce: {
 				ui: {
 					promotions: {
-						currentPage: 12,
-						perPage: 30,
-					},
+						list: {
+							currentPage: 12,
+							perPage: 30,
+						}
+					}
 				},
 				sites: {
 					123: {
