+++ client/extensions/woocommerce/state/ui/promotions/test/list-reducer.js
@@ -8,8 +8,10 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import reducer from '../reducer';
-import { WOOCOMMERCE_PROMOTIONS_PAGE_SET } from 'woocommerce/state/action-types';
+import reducer from '../list-reducer';
+import {
+	WOOCOMMERCE_PROMOTIONS_PAGE_SET,
+} from 'woocommerce/state/action-types';
 
 describe( 'reducer', () => {
 	it( 'should initialize to default values', () => {
